# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ConsumerApp API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/flat": {
            "get": {
                "tags": [
                    "Flat"
                ],
                "description": "Получение списка квартир",
                "responses": {
                    "200": {
                        "description": "Список квартир найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Массив найденных квартир",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Flat"
                                                    }
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Flat"
                ],
                "description": "Добавить квартиру",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Квартира добавлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Flat"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flat/{flat}": {
            "get": {
                "tags": [
                    "Flat"
                ],
                "description": "Получение квартиры",
                "responses": {
                    "200": {
                        "description": "Квартира найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Flat"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Flat"
                ],
                "description": "Отредактировать квартиру",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FlatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Квартира отредактирована",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Flat"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Flat"
                ],
                "description": "Удалить квартиру",
                "responses": {
                    "200": {
                        "description": "Квартира удалёна",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "null"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FlatRequest": {
                "properties": {
                    "title": {
                        "description": "Название квартиры",
                        "type": "string"
                    },
                    "status_id": {
                        "description": "Идентификатор статуса квартиры",
                        "type": "integer"
                    },
                    "full_space": {
                        "description": "Вся площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "floor_count": {
                        "description": "Кол-во этажей",
                        "type": "integer"
                    },
                    "living_space": {
                        "description": "Жилая площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "room_count": {
                        "description": "Кол-во комнат",
                        "type": "integer"
                    },
                    "balconyless_space": {
                        "description": "Площадьь без балконов и т.п.",
                        "type": "number",
                        "format": "float"
                    },
                    "residential_complex_id": {
                        "description": "Идентификатор ЖК, в котором располагается квартира",
                        "type": "integer"
                    },
                    "cost": {
                        "description": "Начальная цена квартиры",
                        "type": "integer"
                    },
                    "is_ready": {
                        "description": "Готова ли квартира",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Flat": {
                "properties": {
                    "id": {
                        "description": "Идентификатор квартиры",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название квартиры",
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/FlatStatus"
                    },
                    "full_space": {
                        "description": "Вся площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "floor_count": {
                        "description": "Кол-во этажей",
                        "type": "integer"
                    },
                    "living_space": {
                        "description": "Жилая площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "room_count": {
                        "description": "Кол-во комнат",
                        "type": "integer"
                    },
                    "balconyless_space": {
                        "description": "Площадьь без балконов и т.п.",
                        "type": "number",
                        "format": "float"
                    },
                    "residential_complex_id": {
                        "description": "Идентификатор ЖК, в котором располагается квартира",
                        "type": "integer"
                    },
                    "cost": {
                        "description": "Начальная цена квартиры",
                        "type": "integer"
                    },
                    "is_ready": {
                        "description": "Готова ли квартира",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "FlatStatus": {
                "properties": {
                    "id": {
                        "description": "Идентификатор статуса квартиры",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название статуса квартиры",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}